error handler fix

+++ backend/utils/errorHandler.js
@@ -1,9 +1,16 @@
+const ApiError = require("./createError");
 const fs = require("fs");
 const errorHandler = (error, req, res, next) => {
     console.log("=========================");
     console.log(error.stack);
     console.log("=========================");
-    const statusCode = error.statusCode || 500;
+    if (error instanceof ApiError) {
+        return res.status(error.statusCode).json({
+            status: false,
+            message: error.message,
+        });
+    }
+    const statusCode = 500;
     if (req.file && req.file.path) {
         fs.unlink(req.file.path, (err) => {
             if (err) console.error("Failed to delete file:", err);
